+++ src/components/Hsi/index.ts
@@ -1,0 +1,1 @@
+export { Hsi } from './Hsi'